Split return-submission data file into single-object and array shapes
returnSubmissionLines is a genuine array of rows, not a legacy array-of-one wrapper, so it moves to its own return-submission-lines.data.js returning the array directly. return-submission.data.js now returns just the bare returnSubmission entity, taking the whole returnLog entity as its parameter.
